requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
apscheduler==3.10.4
    # via -r requirements.in
blinker==1.7.0
    # via
    #   flask
    #   sentry-sdk
cachelib==0.9.0
    # via flask-caching
certifi==2023.11.17
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via flask
controlmyspa==1.0.11
    # via -r requirements.in
flask==3.0.0
    # via
    #   -r requirements.in
    #   flask-caching
    #   sentry-sdk
flask-caching==2.1.0
    # via -r requirements.in
idna==3.6
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   sentry-sdk
    #   werkzeug
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3.post1
    # via apscheduler
requests==2.31.0
    # via
    #   -r requirements.in
    #   controlmyspa
sentry-sdk[flask]==1.39.1
    # via -r requirements.in
six==1.16.0
    # via apscheduler
tzlocal==5.2
    # via apscheduler
urllib3==2.1.0
    # via
    #   requests
    #   sentry-sdk
werkzeug==3.0.1
    # via flask
